Show 'CAL' label when scan is a calibration
To do this, I had to change the beginScan signal in BatchManager
to include a second argument: whether the scan is a calibration.
I also had to add a new pure virtual function to BatchManager
to determin whether the current scan is a calibration. Fortunately
this information was already determined when prepareNextScan exits,
so there was alwyas an appropriate variable containing the needed
information. It may be a good idea to standardize the variable name
and move it to be a protected member of BatchManager. If that is
done, then the additional virtual function can be removed.