Changes to HardwareObject model and initialization
Added a buildCommunication function so that the comm protocol
does not have to be created in the constructor. Instead, the
device indicates a communication type, which is instantiated
after creation but before the initialize function is called.

Also added a d_threaded boolean that allows the hardware to
specify if it should be in its own thread (default is true).

Together, these changes (along with changes to the hardware
manager) allow for easier switching between communication
protocols, and even may allow it to be toggled in the settings
file (eventually).