Initial implementation of amdorbatch
The AMDOR batch acquisition is similar to a DR correlation,
but it attempts to analyze connected networks of lines more
quickly by assigning linkages to trees of related lines
and exploring those trees first.

Currently, it is configured to search for up to 3 linked lines
for a given FT frequency. If finds at least one linkage after
testing all combinations, then the program will test the linked
transition next, again looking for up to 3 child linkages.
This process repeats until the tree is exhausted. Then, the
program picks up where it originally left off.

All new linkages are tested against all previously detected trees
so that a new linkage is not explored thoroughly if it belongs
to a previously explored tree.